feature(p2p): Some cleaning
